requirements.txt
pyautogen
pymemgpt
llama-cpp-python
Pillow